requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.0.0
gspread>=6.0.0
google-auth>=2.25.0

